Add documenting comment
The reason why the previous workflows didn't work is due to the way GitHub Actions handles artifacts and deployments.

When using actions/deploy-pages@v2 or actions/deploy-pages@v3, it expects an artifact to be uploaded using actions/upload-artifact@v3 or actions/upload-pages-artifact@v3. However, in this case, the artifact was not being uploaded correctly, resulting in the "No uploaded artifact was found" error. Used peaceiris/actions-gh-pages@v3 to fix the issue. See the workflow comment.